refactor remote data source module
