+++ data-remote/src/test/kotlin/com/k0d4black/theforce/data/remote/repository/CharacterSearchRepositoryTest.kt
@@ -1,10 +1,9 @@
-package com.k0d4black.theforce.data.repository
+package com.k0d4black.theforce.data.remote.repository
 
 import com.google.common.truth.Truth
-import com.k0d4black.theforce.data.BaseTest
-import com.k0d4black.theforce.data.helpers.EXISTING_SEARCH_PARAMS
-import com.k0d4black.theforce.data.helpers.NON_EXISTENT_SEARCH_PARAMS
-import com.k0d4black.theforce.data.source.CharacterSearchRemoteDataSource
+import com.k0d4black.theforce.data.remote.BaseTest
+import com.k0d4black.theforce.data.remote.helpers.EXISTING_SEARCH_PARAMS
+import com.k0d4black.theforce.data.remote.helpers.NON_EXISTENT_SEARCH_PARAMS
 import com.k0d4black.theforce.domain.repository.ICharacterSearchRepository
 import kotlinx.coroutines.flow.collect
 import kotlinx.coroutines.runBlocking
@@ -18,8 +17,7 @@ internal class CharacterSearchRepositoryTest : BaseTest() {
     @Before
     override fun setup() {
         super.setup()
-        val characterSearchDataSourceMock = CharacterSearchRemoteDataSource(starWarsApiService)
-        characterSearchRepository = CharacterSearchRepository(characterSearchDataSourceMock)
+        characterSearchRepository = CharacterSearchRepository(starWarsApiService)
     }
 
     @Test
