--- data/src/main/kotlin/com/k0d4black/theforce/data/models/entity/SpecieEntity.kt
@@ -1,4 +1,0 @@
-package com.k0d4black.theforce.data.models.entity
-
-
-data class SpecieEntity(val name: String, val language: String)
